+++ src/FuelSystem/FuelTank.cpp
@@ -6,33 +6,33 @@
 
 namespace FuelSystem {
 
-    FuelSystem::FuelTank::FuelTank(int capacity) {
+    FuelSystem::FuelTank::FuelTank(double capacity) {
         this->capacityKg = capacity;
         this->collectorCapacity = 0;
     }
 
-    FuelSystem::FuelTank::FuelTank(int capacity, int collCap) {
+    FuelSystem::FuelTank::FuelTank(double capacity, double collCap) {
         this->capacityKg = capacity;
         this->collectorCapacity = collCap;
     }
 
-    void FuelSystem::FuelTank::setFuel(int fuelKg) {
+    void FuelSystem::FuelTank::setFuel(double fuelKg) {
         this->currFuelKg = fuelKg;
     }
 
-    int FuelTank::getFuel() const {
+    double FuelTank::getFuel() const {
         return this->currFuelKg - this->collectorCapacity;
     }
 
-    int FuelTank::getCollectorFuel() const {
+    double FuelTank::getCollectorFuel() const {
         if (this->currFuelKg > this->collectorCapacity)
             return this->collectorCapacity;
         return this->currFuelKg;
     }
 
-    int FuelTank::addFuel(int amount) {
+    double FuelTank::addFuel(double amount) {
         if (this->currFuelKg + amount > this->capacityKg) { //if overfill, set max and return remainder
-            int temp = this->currFuelKg + amount - this->capacityKg;
+            double temp = this->currFuelKg + amount - this->capacityKg;
             this->currFuelKg = this->capacityKg;
             return temp; // return the fuel we could not place in tank
         }
@@ -41,9 +41,9 @@ namespace FuelSystem {
         return 0;
     }
 
-    int FuelSystem::FuelTank::removeFuel(int amount) {
+    double FuelSystem::FuelTank::removeFuel(double amount) {
         if (this->currFuelKg - this->collectorCapacity - amount < 0) {
-            int temp = this->currFuelKg - this->collectorCapacity;  //temp is the fuel we could remove until tank empty
+            double temp = this->currFuelKg - this->collectorCapacity;  //temp is the fuel we could remove until tank empty
             this->currFuelKg = 0;
             return temp; //return the fuel we could get from the tank
         }
@@ -52,9 +52,9 @@ namespace FuelSystem {
         return amount;
     }
 
-    int FuelSystem::FuelTank::removeFuelCollector(int amount) {
+    double FuelSystem::FuelTank::removeFuelCollector(double amount) {
         if (this->currFuelKg - amount < 0) {
-            int temp = this->currFuelKg;  //temp is the fuel we could remove until tank empty
+            double temp = this->currFuelKg;  //temp is the fuel we could remove until tank empty
             this->currFuelKg = 0;
             return temp; //return the fuel we could get from the tank
         }
